database migration with checks

+++ database/migrations/2017_08_29_143859_make_overwatch_game_if_it_does_not_exist_already.php
@@ -19,14 +19,16 @@ class MakeOverwatchGameIfItDoesNotExistAlready extends Migration
      */
     public function up()
     {
-        $exists = $this->exists();
-        if(!$exists) {
-            $overwatch = new Game();
-            $overwatch->name = $this->name;
-            $overwatch->title = $this->title;
-            $overwatch->uri = $this->uri;
-            $overwatch->description = $this->description;
-            $overwatch->save();
+        if (Schema::connection('mysql_champ')->hasTable('games')) {
+            $exists = $this->exists();
+            if (!$exists) {
+                $overwatch = new Game();
+                $overwatch->name = $this->name;
+                $overwatch->title = $this->title;
+                $overwatch->uri = $this->uri;
+                $overwatch->description = $this->description;
+                $overwatch->save();
+            }
         }
     }
 
@@ -37,11 +39,13 @@ class MakeOverwatchGameIfItDoesNotExistAlready extends Migration
      */
     public function down()
     {
-        $exists = $this->exists();
-        if($exists) {
-            $exists->delete();
-        }
+        if(Schema::connection('mysql_champ')->hasTable('games')) {
+            $exists = $this->exists();
+            if($exists) {
+                $exists->delete();
+            }
 
+        }
     }
 
     /**
